+++ structurizr-dsl/src/test/java/com/structurizr/dsl/DynamicViewContentParserTests.java
@@ -19,7 +19,7 @@ class DynamicViewContentParserTests extends AbstractTests {
             parser.parseRelationship(new DynamicViewDslContext(null), tokens("source", "->", "destination", "description", "technology", "extra"));
             fail();
         } catch (Exception e) {
-            assertEquals("Too many tokens, expected: <identifier> -> <identifier> [description] [technology]", e.getMessage());
+            assertEquals("Too many tokens, expected: [order:] <identifier> -> <identifier> [description] [technology]", e.getMessage());
         }
     }
 
@@ -29,7 +29,7 @@ class DynamicViewContentParserTests extends AbstractTests {
             parser.parseRelationship(new DynamicViewDslContext(null), tokens("source", "->"));
             fail();
         } catch (Exception e) {
-            assertEquals("Expected: <identifier> -> <identifier> [description] [technology]", e.getMessage());
+            assertEquals("Expected: [order:] <identifier> -> <identifier> [description] [technology]", e.getMessage());
         }
     }